requirements.txt
openai_whisper==20230314
pydub==0.25.1
streamlit==1.19.0
